finalização da entrega com erros ja corrigidos.

+++ aplicacoes/gpa-mae/src/main/java/br/ufc/quixada/npi/controller/AlunoController.java
@@ -28,7 +28,7 @@ public class AlunoController {
 	private GenericService<Aluno> genericService;
 	
 	@RequestMapping(value = "{alunoId}", method = RequestMethod.GET)
-	public @ResponseBody Aluno getAlunoJson(@PathVariable("alunoId") Long alunoId) {
+	public @ResponseBody Aluno getAlunoJson(@PathVariable("alunoId") int alunoId) {
 
 		return this.genericService.find(Aluno.class,alunoId);
 
@@ -55,7 +55,7 @@ public class AlunoController {
 	// Metodo Deletar um servidor
 		@RequestMapping(value = "/{alunoId}", method = RequestMethod.DELETE)
 		public @ResponseBody
-		String deletarAluno(@PathVariable("alunoId") Long alunoId) {
+		String deletarAluno(@PathVariable("alunoId") Integer alunoId) {
 			Aluno aluno = this.genericService.find(Aluno.class, alunoId);
 
 			if (aluno == null) {
